Fix a bunch of warnings
I changed my compile command to include -Wall and -Wextra, and I was greeted with a few dozen warnings when I tried to compile.
I fixed many, and ignored many others (e.g. complaints about unused variables and parameters---after all, many functions are still unimplemented). I also fixed some other things.

* Add `return false` statements for bool functions which throw an error before returning false, as a failsafe.

* Fix cout<< implementation.

* Add more cases to some switches.

* Make some `bool` functions `void`. They will throwError() if something goes wrong, anyway.

* Re-const some parameters which I shouldn't have un-const-ed.

* Replace index_t and its variants with the originals, and remove the typedefs.

* Remove useless bounds checking from Screen::forceRenderWorld().

+++ MinePlusPlus/item.cpp
@@ -13,7 +13,7 @@ void Item::pickUp (xcoord_t x, ycoord_t y) {
 }
 
 void Item::spawn (xcoord_t x, ycoord_t y, id_t id) {
-  for (int i = 0; i < (sizeof(items) / sizeof(items[0])); i++) { //find the first NULL pointer in the array and assign it the value of the new item
+  for (uint8_t i = 0; i < (sizeof(items) / sizeof(items[0])); i++) { //find the first NULL pointer in the array and assign it the value of the new item
     if (items[i] == NULL) {
       items[i] = new ItemObject(x, y, id);
       return;
@@ -26,7 +26,7 @@ void Item::spawn (xcoord_t x, ycoord_t y, id_t id) {
 bool Item::despawn () {
   unsigned long oldestTime = millis();
   id_t oldestItemIndex = 255;
-  for (int i = 0; i < (sizeof(items) / sizeof(items[0])); i++) {
+  for (uint8_t i = 0; i < (sizeof(items) / sizeof(items[0])); i++) {
     if (items[i] == NULL) //If the pointer at items[i] is a null pointer, skip over it
       continue;
     if (items[i]->timeOfCreation < oldestTime) { // If a new oldest item is found, set the oldestTime and oldestItemIndex to that new item
